added new services assembleTeam.ts
