+++ src/services/getTeamByIdService.ts
@@ -5,7 +5,7 @@ import {activitiesTable, teamActivitiesTable, teamsTable} from "@/db/schema";
 import { parseZod } from "@/lib/zod";
 import { Team, TeamSchema } from "@/models/Team";
 
-import {BoardSchema} from "@/models/Board";
+import {assembleTeams} from "@/services/assembleTeam";
 
 export const getTeamById = async (id: string): Promise<Team> => {
   const rows = await db
@@ -19,31 +19,7 @@ export const getTeamById = async (id: string): Promise<Team> => {
     throw new Error(`Team with id '${id}' not found`);
   }
 
-  const { teams } = rows[0];
-
-  // Build the board
-  const unSortedBoard = rows.map((row) => ({
-    isCompleted: row.team_activities.completed,
-    points: row.activities.basePoints,
-    activity: row.activities,
-  }));
-
-  const board = unSortedBoard.sort(
-      (a, b) => a.activity.boardOrder - b.activity.boardOrder
-  );
-
-  // Calculate points (e.g. sum of completed activity points)
-  const points = board
-      .filter((entry) => entry.isCompleted)
-      .reduce((sum, entry) => sum + entry.points, 0);
-
-  const team: Team = {
-    id: teams.id,
-    code: teams.code,
-    name: teams.name,
-    points,
-    board: parseZod(BoardSchema, board),
-  };
+  const team = assembleTeams(rows);
 
   return parseZod(TeamSchema, team);
 };
